main program finished, minor fixes required

+++ main.py
@@ -11,7 +11,6 @@ deportes = [Deporte("Fútbol", 10), Deporte("Baloncesto", 15), Deporte("Tenis", 
 
 
 def drop_database():
-
     conn = psycopg2.connect(
         host="localhost",
         database="postgres",
@@ -50,7 +49,7 @@ def create_tables(conn):
     # Create the client table
     cur.execute("""
         CREATE TABLE client (
-            dni SERIAL PRIMARY KEY,
+            dni TEXT PRIMARY KEY,
             nombre TEXT NOT NULL,
             fnac DATE NOT NULL,
             telefono TEXT NOT NULL
@@ -68,7 +67,7 @@ def create_tables(conn):
 
     cur.execute("""
         CREATE TABLE client_sports (
-            dni INTEGER REFERENCES client (dni),
+            dni TEXT REFERENCES client (dni),
             nombre TEXT REFERENCES sports (nombre),
             horario TEXT NOT NULL,
             PRIMARY KEY (dni, nombre)
@@ -100,6 +99,40 @@ conn = psycopg2.connect(
 create_tables(conn)
 
 
+def insertar_deportes():
+    conn = psycopg2.connect(
+        host="localhost",
+        database="polideportivos",
+        user="postgres",
+        password="652050"
+    )
+    conn.autocommit = True
+
+    for deporte in deportes:
+        cur = conn.cursor()
+        cur.execute(
+            "INSERT INTO sports (nombre, precio) VALUES (%s, %s)",
+            (deporte.nombre, deporte.precio_hora)
+        )
+
+    conn.commit()
+
+    conn.close()
+
+
+insertar_deportes()
+
+
+def validar_dni(dni, conn, cur):
+    cur.execute("SELECT dni FROM client WHERE dni = %s", (dni,))
+
+    if cur.rowcount == 0:
+        print("No se ha encontrado ningún cliente con ese DNI.")
+        return False
+
+    return True
+
+
 def alta_cliente():
     name = input("Introduce el nombre completo: ")
     # validacion de dni
@@ -125,8 +158,8 @@ def alta_cliente():
             break
         print("El teléfono debe tener 9 caracteres.")
 
-    # Crear una instancia de la clase Client
-    client = Cliente(name, dni, fnac_formateado, tlf)
+    # Crear una instancia de la clase Cliente
+    client = Cliente(name, dni.upper(), fnac_formateado, tlf)
 
     conn, cur = conexion_bd()
 
@@ -147,14 +180,13 @@ def baja_cliente():
 
     conn, cur = conexion_bd()
 
+    if not validar_dni(dni, conn, cur):
+        return
+
     cur.execute(
         "DELETE FROM client WHERE dni = %s",
-        (dni,)
+        (dni.upper(),)
     )
-
-    if cur.rowcount == 0:
-        print("No se ha encontrado ningún cliente con ese DNI.")
-        return
 
     conn.commit()
 
@@ -163,21 +195,24 @@ def baja_cliente():
     print("Cliente dado de baja correctamente")
 
 
+# TODO - Formatear la fecha de nacimiento a dd/mm/yyyy
 def mostrar_datos_client():
     opcion = input("¿Desea ver los datos de un cliente específico (1) o de todos los clientes (2)? ")
     if opcion == '1':
         dni = input("Ingrese el DNI del cliente: ")
         conn, cur = conexion_bd()
 
+        if not validar_dni(dni, conn, cur):
+            return
+
+
         # Ejecutar una consulta para recuperar los datos del cliente específico
         cur.execute(
             "SELECT dni, nombre, fnac, telefono FROM client WHERE dni = %s",
-            (dni,)
+            (dni.upper(),)
         )
 
-        if cur.rowcount == 0:
-            print("No se ha encontrado ningún cliente con ese DNI.")
-            return
+
 
         cliente = cur.fetchone()
         print("\n")
@@ -212,70 +247,110 @@ def mostrar_datos_client():
 
         conn.commit()
         conn.close()
+    else:
+        print("Opción no válida.")
 
 
 def alta_deporte():
     dni = input("Introduce el DNI del cliente: ")
+    conn, cur = conexion_bd()
+
+    if not validar_dni(dni, conn, cur):
+        return
+
+
     print("Los deportes disponibles son:")
     for index, sport in enumerate(deportes):
         print(index + 1, sport.nombre)
         print("  Precio por hora:", sport.precio_hora)
 
-    sport_index = int(input("Introduce el número del deporte elegido: ")) - 1
-    schedule = input("Introduce el horario elegido: ")
+    index_deporte = int(input("Introduce el número del deporte elegido: ")) - 1
+    horario = input("Introduce el horario elegido: ")
 
-    selected_sport = deportes[sport_index]
-    if not selected_sport:
+    deporte_seleccionado = deportes[index_deporte]
+    if not deporte_seleccionado:
         print("No existe el deporte elegido.")
         return
 
-    selected_sport_name = selected_sport.nombre
-    selected_sport_schedule = selected_sport_name + "-" + schedule
+    nom_deporte = deporte_seleccionado.nombre
+
+    try:
+        conn, cur = conexion_bd()
+
+        cur.execute("INSERT INTO client_sports (dni, nombre, horario) VALUES (%s, %s, %s)",
+                    (dni.upper(), nom_deporte, horario))
+
+        conn.commit()
+        print("Matrícula realizada con éxito.")
+
+    except (Exception, psycopg2.DatabaseError) as error:
+        print("Error al realizar la matrícula:", error)
+
+    finally:
+        if conn is not None:
+            conn.close()
 
 
-def baja_deporte(self):
+def baja_deporte():
+    # validar en la base de datos que el dni existe
     dni = input("Introduce el DNI del cliente: ")
+    conn, cur = conexion_bd()
+
+    if validar_dni(dni, conn, cur) is False:
+        conn.close()
+        print("El cliente con DNI", dni.upper(), "no está matriculado en ningún deporte.")
+        return
+
     sport_name = input("Introduce el nombre del deporte: ")
 
     conn, cur = conexion_bd()
 
     cur.execute(
-        "SELECT sports FROM client WHERE dni = %s",
-        (dni,)
-    )
-    enrolled_sports = cur.fetchone()[0]
-
-    if enrolled_sports is None or sport_name not in enrolled_sports:
-        print("El cliente no está matriculado en este deporte.")
-        return
-
-    enrolled_sports.remove(sport_name)
-
-    cur.execute(
-        "UPDATE client SET sports = %s WHERE dni = %s",
-        (enrolled_sports, dni)
+        "DELETE FROM client_sports WHERE nombre = %s AND dni = %s",
+        (sport_name, dni.upper())
     )
     conn.commit()
     conn.close()
     print("Se ha eliminado el deporte con éxito.")
 
 
-# TODO - Falta Arreglar
-def mostrar_deport_client(self):
+def mostrar_deportes_cliente():
+    # validar en la base de datos que el dni existe
+
     dni = input("Introduce el DNI del cliente: ")
 
     conn, cur = conexion_bd()
-    cur.execute("SELECT sport FROM client WHERE dni = %s", (dni,))
-    enrolled_sports = cur.fetchone()[0]
 
-    if not enrolled_sports:
-        print("El cliente no está matriculado en ningún deporte.")
+    cur.execute(
+        "SELECT dni FROM client WHERE dni = %s",
+        (dni.upper(),)
+    )
+
+    if cur.rowcount == 0:
+        print("No se ha encontrado ningún cliente con ese DNI.")
         return
 
-    print("El cliente está matriculado en los siguientes deportes:")
+    conn.commit()
+    conn.close()
+
+    conn, cur = conexion_bd()
+    cur.execute("SELECT nombre, horario FROM client_sports WHERE dni = %s", (dni,))
+    enrolled_sports = cur.fetchall()
+
+    # conseguir el nombre del dni del cliente
+    cur.execute("SELECT nombre FROM client WHERE dni = %s", (dni.upper(),))
+    nombre = cur.fetchone()[0]
+
+    if not enrolled_sports:
+        print("El cliente " + nombre + "no está matriculado en ningún deporte.")
+        return
+
+    print("El cliente " + nombre + " está matriculado en los siguientes deportes:")
     for sport in enrolled_sports:
-        sport_name, schedule = sport.split("-")
-        print("- Deporte:", sport_name)
+        print("\n")
+        print("  Deporte:", sport[0])
+        print("  Horario:", sport[1])
+        print("\n")
 
     conn.close()
 
@@ -301,12 +376,11 @@ def menu():
         3: mostrar_datos_client,
         4: alta_deporte,
         5: baja_deporte,
-        6: mostrar_deport_client,
+        6: mostrar_deportes_cliente,
         7: exit
     }
 
-    func = switcher.get(opcion, lambda: "Opción no válida")
-    func()
+    switcher.get(opcion, lambda: print("Opción no válida."))()
 
 
 # Función reutilizable para conectar con la base de datos
